Global Tracking Workflow: Move all reader specs to readers library, to reduce the dependency chain of its users, and avoid cyclic dependencies
