+++ Detectors/GlobalTrackingWorkflow/CMakeLists.txt
@@ -14,22 +14,19 @@ o2_add_library(GlobalTrackingWorkflow
                SOURCES src/TrackWriterTPCITSSpec.cxx
                        src/TPCITSMatchingSpec.cxx
                        src/MatchTPCITSWorkflow.cxx
-                       src/TrackTPCITSReaderSpec.cxx
                        src/PrimaryVertexingSpec.cxx
                        src/PrimaryVertexWriterSpec.cxx
-                       src/PrimaryVertexReaderSpec.cxx
                        src/VertexTrackMatcherSpec.cxx
                        src/SecondaryVertexingSpec.cxx
                        src/SecondaryVertexWriterSpec.cxx
-                       src/SecondaryVertexReaderSpec.cxx
                        src/CosmicsMatchingSpec.cxx
-		       src/TrackCosmicsWriterSpec.cxx
-		       src/TrackCosmicsReaderSpec.cxx		       
+                       src/TrackCosmicsWriterSpec.cxx
                PUBLIC_LINK_LIBRARIES O2::GlobalTracking
+                                     O2::GlobalTrackingWorkflowReaders
                                      O2::ITStracking
                                      O2::ITSWorkflow
                                      O2::TPCWorkflow
-                                     O2::FT0Workflow                                     
+                                     O2::FT0Workflow
                                      O2::ITSMFTWorkflow
                                      O2::SimulationDataFormat
                                      O2::DetectorsVertexing)
@@ -68,3 +65,4 @@ o2_add_executable(vertexing-workflow
 add_subdirectory(tofworkflow)
 add_subdirectory(tpcinterpolationworkflow)
 add_subdirectory(helpers)
+add_subdirectory(readers)
