+++ Detectors/GlobalTrackingWorkflow/helpers/CMakeLists.txt
@@ -12,7 +12,6 @@ o2_add_library(GlobalTrackingWorkflowHelpers
                SOURCES src/InputHelper.cxx
                PUBLIC_LINK_LIBRARIES O2::ReconstructionDataFormats
                PRIVATE_LINK_LIBRARIES O2::GlobalTracking # We link all the workflow libs etc as PRIVATE, such that we do not expose tons of headers automatically
-                                      O2::GlobalTrackingWorkflow
                                       O2::ITSWorkflow
                                       O2::ITSMFTWorkflow
                                       O2::TPCWorkflow
