+++ Detectors/GlobalTrackingWorkflow/tofworkflow/CMakeLists.txt
@@ -15,8 +15,8 @@ o2_add_library(TOFWorkflow
                        src/TOFCalibWriterSpec.cxx
                        src/TOFMatchedReaderSpec.cxx
                        src/CalibInfoReaderSpec.cxx
-               PUBLIC_LINK_LIBRARIES O2::Framework O2::TOFBase O2::DataFormatsTOF O2::TOFReconstruction 
-                                     O2::GlobalTracking O2::GlobalTrackingWorkflow O2::TOFWorkflowUtils
+               PUBLIC_LINK_LIBRARIES O2::Framework O2::TOFBase O2::DataFormatsTOF O2::TOFReconstruction
+                                     O2::GlobalTracking O2::GlobalTrackingWorkflowReaders O2::TOFWorkflowUtils
                                      O2::TOFCalibration O2::DataFormatsFT0 O2::FT0Reconstruction O2::FT0Workflow
                                      O2::TPCWorkflow)
